[Rename] `testUtility.ts` to `utility.ts`
